3_F_jak_Ficzur/praca_domowa/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
jinja2
